// shared/types/personnel.types.ts
export interface VitalSigns {
  bloodPressure?: string;
  heartRate?: string;
  respiratoryRate?: string;
  temperature?: string;
  spO2?: string;
  weight?: string;
  height?: string;
}
// ...
export interface CreateClinicalRecordDTO {
  date: Date | string;
  assessmentDateTime?: string;
  illnessOnsetDateTime?: string;
  gender?: string;
  complaint: string;
  diagnosis: string;
  treatment: string;
  physician: string;
  vitalSigns?: VitalSigns;
  physicalExam?: string;
  pmhx?: string;
  psHx?: string;
  fhx?: string;
  shx?: string;
  allergies?: string;
  medications?: string;
  covidVaccinated?: boolean;
  covidVaccineDetails?: string;
  notes?: string;
}